[feature/azure_blob_storage] Adds support for storing file uploads as azure blobs
Windows azure is a clould platform by microsoft. This patch allows the use of azure blobs for storing uploads. This is required for persisting uploads and sharing them between instances.

Note that there is no plan for merging this into the mainline release. Depolyment requires the addition of the azure php API and the creation of a deployment package. We are preparing documentation for the required steps.