[feature/9550] General "Easy" Hooks Introduces a basic hook system into the develop branch. Can be tested by calling "$hook_handler->register_mod('test_mod')" .
Names, paths, and the schema are subject to change and largely depend on the yet-to-be-designed MOD installer.
Bridging existing plugin interfaces into the new MOD concept is a major challenge in this regard.

This code is largely untested and has bugs. Please report any you find at area51.

The placement of hook invokations in the core code is the next phase.