[bug/9698] Avoid PHP 5.3 deprecated messages
phpBB3 strives to work with php4 as well as php5 including 5.3. However, the use of OOP in
php4 requires constructs causing warnings under php5.3. This commit is intended to remove the display of such warnings
when the files are visited directly, i.e. without phpBB being able to alter error reporting.